Add about section
To add a new section just add another category folder
under content. To have it show up in the header you
need to add another section to `config.toml`.
Generally files in such a directory/category get
listed.
The file `_index.md` is special and contains content
for the "listing", or in this case for the about
page.